+++ blocklets/ai-studio/api/src/routes/ai.ts
@@ -1,6 +1,12 @@
+import { defaultImageModel, getSupportedImagesModels } from '@app/libs/common';
 import { chatCompletions, imageGenerations, proxyToAIKit } from '@blocklet/ai-kit/api/call';
-import { CallAI, GetAssistant, nextTaskId, runAssistant } from '@blocklet/ai-runtime/core';
-import { AssistantResponseType, RunAssistantResponse, isPromptAssistant } from '@blocklet/ai-runtime/types';
+import { CallAI, CallAIImage, GetAssistant, nextTaskId, runAssistant } from '@blocklet/ai-runtime/core';
+import {
+  AssistantResponseType,
+  RunAssistantResponse,
+  isImageAssistant,
+  isPromptAssistant,
+} from '@blocklet/ai-runtime/types';
 import user from '@blocklet/sdk/lib/middlewares/user';
 import compression from 'compression';
 import { Router } from 'express';
@@ -51,19 +57,56 @@ router.post('/call', user(), compression(), ensureComponentCallOrAuth(), async (
 
   const repository = await getRepository({ projectId: input.projectId });
 
-  const callAI: CallAI = ({ assistant, input }) => {
+  const callAI: CallAI = async ({ assistant, input, outputModel = false }) => {
     const promptAssistant = isPromptAssistant(assistant) ? assistant : undefined;
 
-    return chatCompletions({
-      ...input,
+    const model = {
       model: input.model || promptAssistant?.model || project.model || defaultModel,
       temperature: input.temperature ?? promptAssistant?.temperature ?? project.temperature,
       topP: input.topP ?? promptAssistant?.topP ?? project.topP,
       presencePenalty: input.presencePenalty ?? promptAssistant?.presencePenalty ?? project.presencePenalty,
       frequencyPenalty: input.frequencyPenalty ?? promptAssistant?.frequencyPenalty ?? project.frequencyPenalty,
+    };
+    const chatCompletionChunk = await chatCompletions({
+      ...input,
+      ...model,
       // FIXME: should be maxTokens - prompt tokens
       // maxTokens: input.maxTokens ?? project.maxTokens,
     });
+
+    if (outputModel) {
+      return {
+        chatCompletionChunk,
+        modelInfo: model,
+      };
+    }
+    return chatCompletionChunk as any;
+  };
+
+  const callAIImage: CallAIImage = async ({ assistant, input, outputModel = false }) => {
+    const imageAssistant = isImageAssistant(assistant) ? assistant : undefined;
+    const supportImages = await getSupportedImagesModels();
+    const imageModel = supportImages.find((i) => i.model === (imageAssistant?.model || defaultImageModel));
+
+    const model = {
+      model: input.model || imageModel?.model,
+      n: input.n || imageModel?.nDefault,
+      quality: input.quality || imageModel?.qualityDefault,
+      style: input.style || imageModel?.styleDefault,
+      size: input.size || imageModel?.sizeDefault,
+    };
+    const imageRes = await imageGenerations({
+      ...input,
+      ...model,
+    });
+
+    if (outputModel) {
+      return {
+        imageRes,
+        modelInfo: model,
+      };
+    }
+    return imageRes as any;
   };
 
   const getAssistant: GetAssistant = (fileId: string, options) => {
@@ -105,7 +148,7 @@ router.post('/call', user(), compression(), ensureComponentCallOrAuth(), async (
 
     const result = await runAssistant({
       callAI,
-      callAIImage: ({ input }) => imageGenerations(input),
+      callAIImage,
       taskId,
       getAssistant,
       assistant,