+++ packages/ai-runtime/src/core/assistant/index.ts
@@ -30,25 +30,59 @@ import {
   isFunctionAssistant,
   isPromptAssistant,
 } from '../../types';
-import { ImageAssistant, Mustache, Role, User, isImageAssistant } from '../../types/assistant';
+import { ImageAssistant, Mustache, OnTaskCompletion, Role, User, isImageAssistant } from '../../types/assistant';
 import { AssistantResponseType, ExecutionPhase, RunAssistantResponse } from '../../types/runtime';
 
 export type RunAssistantCallback = (e: RunAssistantResponse) => void;
+
+export class ToolCompletionDirective extends Error {
+  type: OnTaskCompletion;
+
+  constructor(message: string, type: OnTaskCompletion) {
+    super(message);
+    this.type = type;
+  }
+}
 
 export interface GetAssistant {
   (assistantId: string, options: { rejectOnEmpty: true | Error }): Promise<Assistant>;
   (assistantId: string, options?: { rejectOnEmpty?: false }): Promise<Assistant | null>;
 }
 
-export type CallAI = (options: {
+export type Options = {
   assistant: Assistant;
   input: ChatCompletionInput;
-}) => Promise<ReadableStream<ChatCompletionChunk>>;
+};
 
-export type CallAIImage = (options: {
+export type ImageOptions = {
   assistant: Assistant;
   input: ImageGenerationInput;
-}) => Promise<ImageGenerationResponse>;
+};
+
+export type ModelInfo = {
+  model: string;
+  temperature?: number;
+  topP?: number;
+  presencePenalty?: number;
+  frequencyPenalty?: number;
+};
+
+export interface CallAI {
+  (options: Options & { outputModel: true }): Promise<{
+    modelInfo: ModelInfo;
+    chatCompletionChunk: ReadableStream<ChatCompletionChunk>;
+  }>;
+  (options: Options & { outputModel?: false }): Promise<ReadableStream<ChatCompletionChunk>>;
+  (options: Options & { outputModel: boolean }): any;
+}
+
+export interface CallAIImage {
+  (options: ImageOptions & { outputModel: true }): Promise<{
+    modelInfo: ModelInfo;
+    imageRes: ImageGenerationResponse;
+  }>;
+  (options: ImageOptions & { outputModel?: false }): Promise<ImageGenerationResponse>;
+}
 
 const taskIdGenerator = new Worker();
 
@@ -61,6 +95,7 @@ export async function runAssistant({
   getAssistant,
   assistant,
   parameters = {},
+  parentTaskId,
   callback,
   user,
 }: {
@@ -70,59 +105,89 @@ export async function runAssistant({
   getAssistant: GetAssistant;
   assistant: Assistant;
   parameters?: { [key: string]: any };
+  parentTaskId?: string;
   callback?: RunAssistantCallback;
   user?: User;
 }): Promise<any> {
-  if (isPromptAssistant(assistant)) {
-    return runPromptAssistant({
-      taskId,
-      callAI,
-      callAIImage,
-      getAssistant,
-      assistant,
-      parameters,
-      callback,
-      user,
-    });
-  }
+  callback?.({
+    type: AssistantResponseType.EXECUTE,
+    taskId,
+    parentTaskId,
+    assistantId: assistant.id,
+    assistantName: assistant.name,
+    execution: { currentPhase: ExecutionPhase.EXECUTE_ASSISTANT_START },
+  });
+  try {
+    if (isPromptAssistant(assistant)) {
+      return await runPromptAssistant({
+        taskId,
+        parentTaskId,
+        callAI,
+        callAIImage,
+        getAssistant,
+        assistant,
+        parameters,
+        callback,
+        user,
+      });
+    }
 
-  if (isImageAssistant(assistant)) {
-    return runImageAssistant({
-      taskId,
-      callAI,
-      callAIImage,
-      getAssistant,
-      assistant,
-      parameters,
-      callback,
-      user,
-    });
-  }
+    if (isImageAssistant(assistant)) {
+      return await runImageAssistant({
+        taskId,
+        parentTaskId,
+        callAI,
+        callAIImage,
+        getAssistant,
+        assistant,
+        parameters,
+        callback,
+        user,
+      });
+    }
 
-  if (isFunctionAssistant(assistant)) {
-    return runFunctionAssistant({
-      getAssistant,
-      callAI,
-      callAIImage,
-      taskId,
-      assistant,
-      parameters,
-      callback,
-      user,
-    });
-  }
+    if (isFunctionAssistant(assistant)) {
+      return await runFunctionAssistant({
+        getAssistant,
+        callAI,
+        callAIImage,
+        taskId,
+        parentTaskId,
+        assistant,
+        parameters,
+        callback,
+        user,
+      });
+    }
 
-  if (isApiAssistant(assistant)) {
-    return runApiAssistant({
-      getAssistant,
-      callAI,
-      callAIImage,
-      taskId,
-      assistant,
-      parameters,
-      callback,
-      user,
-    });
+    if (isApiAssistant(assistant)) {
+      return await runApiAssistant({
+        getAssistant,
+        callAI,
+        callAIImage,
+        taskId,
+        parentTaskId,
+        assistant,
+        parameters,
+        callback,
+        user,
+      });
+    }
+  } catch (e) {
+    if (e instanceof ToolCompletionDirective) {
+      if (e.type === OnTaskCompletion.EXIT) {
+        callback?.({
+          type: AssistantResponseType.EXECUTE,
+          taskId,
+          parentTaskId,
+          assistantId: assistant.id,
+          assistantName: assistant.name,
+          execution: { currentPhase: ExecutionPhase.EXECUTE_SELECT_STOP },
+        });
+        return undefined;
+      }
+    }
+    throw e;
   }
 
   throw new Error('Unimplemented');
@@ -136,6 +201,7 @@ async function runFunctionAssistant({
   assistant,
   context,
   parameters,
+  parentTaskId,
   callback,
   user,
 }: {
@@ -146,6 +212,7 @@ async function runFunctionAssistant({
   assistant: FunctionAssistant;
   context?: { [key: string]: any };
   parameters?: { [key: string]: any };
+  parentTaskId?: string;
   callback?: RunAssistantCallback;
   user?: User;
 }) {
@@ -157,6 +224,7 @@ async function runFunctionAssistant({
         getAssistant,
         parameters,
         executeBlocks: assistant.prepareExecutes,
+        parentTaskId: taskId,
         callback,
         user,
       })
@@ -222,9 +290,20 @@ async function runFunctionAssistant({
   callback?.({
     type: AssistantResponseType.EXECUTE,
     taskId,
+    parentTaskId,
     assistantId: assistant.id,
     assistantName: assistant.name,
-    execution: { currentPhase: ExecutionPhase.EXECUTE_ASSISTANT_START },
+    execution: { currentPhase: ExecutionPhase.EXECUTE_ASSISTANT_RUNNING },
+  });
+
+  callback?.({
+    type: AssistantResponseType.INPUT,
+    assistantId: assistant.id,
+    taskId,
+    parentTaskId,
+    assistantName: assistant.name,
+    inputParameters: parameters,
+    fnArgs: args,
   });
 
   const result = await module.default(args);
@@ -234,6 +313,15 @@ async function runFunctionAssistant({
     taskId,
     assistantId: assistant.id,
     delta: { content: typeof result === 'string' ? result : JSON.stringify(result) },
+  });
+
+  callback?.({
+    type: AssistantResponseType.EXECUTE,
+    taskId,
+    parentTaskId,
+    assistantId: assistant.id,
+    assistantName: assistant.name,
+    execution: { currentPhase: ExecutionPhase.EXECUTE_ASSISTANT_END },
   });
 
   return result;
@@ -246,6 +334,7 @@ async function runApiAssistant({
   taskId,
   assistant,
   parameters,
+  parentTaskId,
   callback,
   user,
 }: {
@@ -255,6 +344,7 @@ async function runApiAssistant({
   taskId: string;
   assistant: ApiAssistant;
   parameters?: { [key: string]: any };
+  parentTaskId?: string;
   callback?: RunAssistantCallback;
   user?: User;
 }) {
@@ -268,6 +358,7 @@ async function runApiAssistant({
         getAssistant,
         parameters,
         executeBlocks: assistant.prepareExecutes,
+        parentTaskId: taskId,
         callback,
         user,
       })
@@ -291,9 +382,20 @@ async function runApiAssistant({
   callback?.({
     type: AssistantResponseType.EXECUTE,
     taskId,
+    parentTaskId,
     assistantId: assistant.id,
     assistantName: assistant.name,
-    execution: { currentPhase: ExecutionPhase.EXECUTE_ASSISTANT_START },
+    execution: { currentPhase: ExecutionPhase.EXECUTE_ASSISTANT_RUNNING },
+  });
+
+  callback?.({
+    type: AssistantResponseType.INPUT,
+    assistantId: assistant.id,
+    taskId,
+    assistantName: assistant.name,
+    parentTaskId,
+    inputParameters: parameters,
+    apiArgs: args,
   });
 
   const method = assistant.requestMethod || 'GET';
@@ -327,6 +429,15 @@ async function runApiAssistant({
     delta: { content: typeof result === 'string' ? result : JSON.stringify(result || error) },
   });
 
+  callback?.({
+    type: AssistantResponseType.EXECUTE,
+    taskId,
+    parentTaskId,
+    assistantId: assistant.id,
+    assistantName: assistant.name,
+    execution: { currentPhase: ExecutionPhase.EXECUTE_ASSISTANT_END },
+  });
+
   return result;
 }
 
@@ -343,6 +454,7 @@ async function runPromptAssistant({
   getAssistant,
   assistant,
   parameters,
+  parentTaskId,
   callback,
   user,
 }: {
@@ -352,6 +464,7 @@ async function runPromptAssistant({
   getAssistant: GetAssistant;
   assistant: PromptAssistant;
   parameters: { [key: string]: any };
+  parentTaskId?: string;
   callback?: RunAssistantCallback;
   user?: User;
 }) {
@@ -368,6 +481,7 @@ async function runPromptAssistant({
     getAssistant,
     executeBlocks,
     parameters,
+    parentTaskId: taskId,
     callback,
     user,
   });
@@ -427,17 +541,28 @@ async function runPromptAssistant({
     .flat()
     .filter((i): i is Required<NonNullable<typeof i>> => !!i?.content);
 
-  callback?.({ type: AssistantResponseType.INPUT, taskId, assistantId: assistant.id, input: { messages } });
   callback?.({
     type: AssistantResponseType.EXECUTE,
     taskId,
+    parentTaskId,
     assistantId: assistant.id,
     assistantName: assistant.name,
-    execution: { currentPhase: ExecutionPhase.EXECUTE_ASSISTANT_START },
+    execution: { currentPhase: ExecutionPhase.EXECUTE_ASSISTANT_RUNNING },
+  });
+
+  callback?.({
+    type: AssistantResponseType.INPUT,
+    assistantId: assistant.id,
+    parentTaskId,
+    taskId,
+    assistantName: assistant.name,
+    inputParameters: parameters,
+    promptMessages: messages,
   });
 
   const res = await callAI({
     assistant,
+    outputModel: true,
     input: {
       stream: true,
       messages,
@@ -451,15 +576,39 @@ async function runPromptAssistant({
 
   let result = '';
 
-  for await (const chunk of res) {
+  for await (const chunk of res.chatCompletionChunk) {
+    result += chunk.delta.content || '';
     callback?.({
       type: AssistantResponseType.CHUNK,
       taskId,
       assistantId: assistant.id,
       delta: { content: chunk.delta.content },
     });
-    result += chunk.delta.content || '';
   }
+
+  callback?.({
+    type: AssistantResponseType.INPUT,
+    assistantId: assistant.id,
+    ...(parentTaskId
+      ? {
+          parentTaskId,
+          modelParameters: res.modelInfo,
+        }
+      : { parentTaskId }),
+    taskId,
+    assistantName: assistant.name,
+    inputParameters: parameters,
+    promptMessages: messages,
+  });
+
+  callback?.({
+    type: AssistantResponseType.EXECUTE,
+    taskId,
+    parentTaskId,
+    assistantId: assistant.id,
+    assistantName: assistant.name,
+    execution: { currentPhase: ExecutionPhase.EXECUTE_ASSISTANT_END },
+  });
 
   return result;
 }
@@ -471,6 +620,7 @@ async function runImageAssistant({
   getAssistant,
   assistant,
   parameters,
+  parentTaskId,
   callback,
   user,
 }: {
@@ -480,6 +630,7 @@ async function runImageAssistant({
   getAssistant: GetAssistant;
   assistant: ImageAssistant;
   parameters: { [key: string]: any };
+  parentTaskId?: string;
   callback?: RunAssistantCallback;
   user?: User;
 }) {
@@ -493,6 +644,7 @@ async function runImageAssistant({
         getAssistant,
         parameters,
         executeBlocks: assistant.prepareExecutes,
+        parentTaskId: taskId,
         callback,
         user,
       })
@@ -518,13 +670,27 @@ async function runImageAssistant({
   callback?.({
     type: AssistantResponseType.EXECUTE,
     taskId,
+    parentTaskId,
     assistantId: assistant.id,
     assistantName: assistant.name,
-    execution: { currentPhase: ExecutionPhase.EXECUTE_ASSISTANT_START },
+    execution: { currentPhase: ExecutionPhase.EXECUTE_ASSISTANT_RUNNING },
   });
 
-  const { data } = await callAIImage({
+  callback?.({
+    type: AssistantResponseType.INPUT,
+    assistantId: assistant.id,
+    taskId,
+    parentTaskId,
+    assistantName: assistant.name,
+    inputParameters: parameters,
+  });
+
+  const {
+    imageRes: { data },
+    modelInfo,
+  } = await callAIImage({
     assistant,
+    outputModel: true,
     input: {
       prompt,
       n: assistant.n,
@@ -536,7 +702,29 @@ async function runImageAssistant({
     },
   });
 
+  callback?.({
+    type: AssistantResponseType.INPUT,
+    assistantId: assistant.id,
+    ...(parentTaskId
+      ? {
+          parentTaskId,
+          modelParameters: modelInfo,
+        }
+      : { parentTaskId }),
+    taskId,
+    assistantName: assistant.name,
+  });
+
   callback?.({ type: AssistantResponseType.CHUNK, taskId, assistantId: assistant.id, delta: { images: data } });
+
+  callback?.({
+    type: AssistantResponseType.EXECUTE,
+    taskId,
+    parentTaskId,
+    assistantId: assistant.id,
+    assistantName: assistant.name,
+    execution: { currentPhase: ExecutionPhase.EXECUTE_ASSISTANT_END },
+  });
 
   return data;
 }
@@ -548,6 +736,7 @@ async function runExecuteBlocks({
   getAssistant,
   parameters,
   executeBlocks,
+  parentTaskId,
   callback,
   user,
 }: {
@@ -557,6 +746,7 @@ async function runExecuteBlocks({
   getAssistant: GetAssistant;
   parameters?: { [key: string]: any };
   executeBlocks: ExecuteBlock[];
+  parentTaskId?: string;
   callback?: RunAssistantCallback;
   user?: User;
 }) {
@@ -579,6 +769,7 @@ async function runExecuteBlocks({
         executeBlock,
         parameters: variables,
         datasets,
+        parentTaskId,
         callback,
         user,
       });
@@ -605,6 +796,7 @@ async function runExecuteBlock({
   executeBlock,
   parameters,
   datasets,
+  parentTaskId,
   callback,
   user,
 }: {
@@ -616,6 +808,7 @@ async function runExecuteBlock({
   executeBlock: ExecuteBlock;
   parameters?: { [key: string]: any };
   datasets?: DatasetObject[];
+  parentTaskId?: string;
   callback?: RunAssistantCallback;
   user?: User;
 }) {
@@ -625,6 +818,7 @@ async function runExecuteBlock({
   callback?.({
     type: AssistantResponseType.EXECUTE,
     taskId,
+    parentTaskId,
     assistantId: assistant.id,
     assistantName: assistant.name,
     execution: {
@@ -690,6 +884,7 @@ async function runExecuteBlock({
             getAssistant,
             assistant: toolAssistant,
             parameters: args,
+            parentTaskId,
             callback,
           });
         })
@@ -757,6 +952,14 @@ async function runExecuteBlock({
       )
     ).filter((i): i is NonNullable<typeof i> => !isNil(i));
 
+    callback?.({
+      type: AssistantResponseType.INPUT,
+      assistantId: assistant.id,
+      parentTaskId,
+      taskId,
+      modelParameters: executeBlock.executeModel,
+      assistantName: `${executeBlock.variable ?? assistant.name}-select`,
+    });
     const response = await callAI({
       assistant,
       input: {
@@ -781,7 +984,7 @@ async function runExecuteBlock({
           type: AssistantResponseType.CHUNK,
           taskId,
           assistantId: assistant.id,
-          delta: { content: chunk.delta.content },
+          delta: { content: chunk.delta.content || '' },
         });
       }
 
@@ -803,7 +1006,6 @@ async function runExecuteBlock({
     }
 
     const toolAssistantMap = Object.fromEntries(toolAssistants.map((i) => [i.function.name, i]));
-
     const result =
       calls &&
       (await Promise.all(
@@ -848,15 +1050,22 @@ async function runExecuteBlock({
             }) ?? []
           );
 
-          return runAssistant({
+          const res = await runAssistant({
             taskId: taskIdGenerator.nextId().toString(),
             callAI,
             callAIImage,
             getAssistant,
             assistant: toolAssistant,
             parameters: args,
+            parentTaskId: taskId,
             callback,
           });
+
+          if (tool.tool?.onEnd === OnTaskCompletion.EXIT) {
+            throw new ToolCompletionDirective('The task has been stop. The tool will now exit.', OnTaskCompletion.EXIT);
+          }
+
+          return res;
         })
       ));