+++ packages/ai-runtime/src/types/runtime/index.ts
@@ -4,7 +4,10 @@ import { Role } from '../assistant';
 
 export enum ExecutionPhase {
   EXECUTE_BLOCK_START = 'EXECUTE_BLOCK_START',
+  EXECUTE_SELECT_STOP = 'EXECUTE_SELECT_STOP',
   EXECUTE_ASSISTANT_START = 'EXECUTE_ASSISTANT_START',
+  EXECUTE_ASSISTANT_RUNNING = 'EXECUTE_BLOCK_RUNNING',
+  EXECUTE_ASSISTANT_END = 'EXECUTE_ASSISTANT_END',
 }
 export enum AssistantResponseType {
   ERROR = 'ERROR',
@@ -14,12 +17,10 @@ export enum AssistantResponseType {
   EXECUTE = 'EXECUTE',
 }
 
-export type InputMessages = {
-  messages: Array<{
-    role: Role;
-    content: string;
-  }>;
-};
+export type PromptMessages = Array<{
+  role: Role;
+  content: string;
+}>;
 
 export type RunAssistantResponse =
   | RunAssistantChunk
@@ -31,13 +32,30 @@ export type RunAssistantResponse =
 export type RunAssistantInput = {
   type: AssistantResponseType.INPUT;
   taskId: string;
+  parentTaskId?: string;
   assistantId: string;
-  input: InputMessages;
+  assistantName?: string;
+  inputParameters?: { [key: string]: string };
+  apiArgs?: any;
+  fnArgs?: any;
+  promptMessages?: PromptMessages;
+  modelParameters?: {
+    temperature?: number;
+    topP?: number;
+    presencePenalty?: number;
+    frequencyPenalty?: number;
+    model?: string;
+    n?: number;
+    quality?: string;
+    style?: string;
+    size?: string;
+  };
 };
 
 export type RunAssistantExecute = {
   type: AssistantResponseType.EXECUTE;
   taskId: string;
+  parentTaskId?: string;
   assistantId: string;
   assistantName?: string;
   execution:
@@ -48,6 +66,15 @@ export type RunAssistantExecute = {
       }
     | {
         currentPhase: ExecutionPhase.EXECUTE_ASSISTANT_START;
+      }
+    | {
+        currentPhase: ExecutionPhase.EXECUTE_SELECT_STOP;
+      }
+    | {
+        currentPhase: ExecutionPhase.EXECUTE_ASSISTANT_END;
+      }
+    | {
+        currentPhase: ExecutionPhase.EXECUTE_ASSISTANT_RUNNING;
       };
 };